Bug 1159506 - Make GC events use TimeStamp. r=terrence
UltraBlame original commit: 35091aff4234a3972cbc3b1cbe0de5f239f84cdf

+++ js/public/GCAPI.h
@@ -285,8 +285,8 @@ class GarbageCollectionEvent
     
     
     struct Collection {
-        int64_t startTimestamp;
-        int64_t endTimestamp;
+        double startTimestamp;
+        double endTimestamp;
     };
